Extract query strategies to separate file
